04-01 removed the aborted event of a request stream

+++ 03-streams/04-file-server-post/server.js
@@ -31,7 +31,10 @@ server.on('request', (req, res) => {
         fileStream.destroy(error);
       });
 
-      req.on('aborted', () => {
+      req.on('close', () => {
+        if (req.readableEnded) return;
+        // req.readableEnded is false for the aborded stream
+
         fileStream.end();
 
         fs.unlink(filepath, () => {});
